refactor(napcat): 模块内分层对齐 domain/application/infra/http
napcat 是唯一仍在用 service/ 目录的模块，结构落后于 metric/scheduler 等
金标准。本次只做目录迁移与 import 路径重写，无任何逻辑改动：

- service/ → application/（消灭唯一的 service/ 异类目录；gateway port 与
  契约类型保持同层，对齐 auth/llm 把契约类型贴着 port 放的现状）
- dao/ → infra/（dao 接口入 infra/、实现入 infra/impl/，对齐 metric）
- schema/napcat-segment.ts → domain/napcat-segment.ts（napcat 消息数据契约归 domain）

纯文件移动 + 路径修正，不触碰 system prompt、InvokeTool 子工具集与任何进入
Agent 消息的序列化，对在飞会话的 KV 前缀零影响。

build / typecheck / test(619) / lint(0 error) / format 全绿。
