+++ apps/server/src/agent/apps/qq/tools/view-forward.tool.ts
@@ -1,6 +1,6 @@
 import { z } from "zod";
 import { ZodToolComponent, type ToolExecutionResult, type ToolKind } from "@kagami/agent-runtime";
-import { FORWARD_ID_DISPLAY_PREFIX } from "../../../../napcat/service/napcat-gateway/shared.js";
+import { FORWARD_ID_DISPLAY_PREFIX } from "../../../../napcat/application/napcat-gateway/shared.js";
 import type { QqApp } from "../qq.app.js";
 
 // forward_id 容忍 string | number：占位符带 fwd- 前缀本会强制成字符串，但 LLM 偶尔仍会把它
