+++ apps/server/src/app/agent-runtime.factory.ts
@@ -18,9 +18,9 @@ import { DefaultLlmPlaygroundService } from "../llm/application/llm-playground.i
 import type { LlmPlaygroundService } from "../llm/application/llm-playground.service.js";
 import type { MetricService } from "../metric/application/metric.service.js";
 import type { ConfigManager } from "../config/config.manager.js";
-import type { NapcatQqMessageDao } from "../napcat/dao/napcat-group-message.dao.js";
-import type { NapcatGatewayPersistenceWriter } from "../napcat/service/napcat-gateway/event-persistence-writer.js";
-import type { NapcatImageMessageAnalyzer } from "../napcat/service/napcat-gateway/image-message-analyzer.js";
+import type { NapcatQqMessageDao } from "../napcat/infra/napcat-group-message.dao.js";
+import type { NapcatGatewayPersistenceWriter } from "../napcat/application/napcat-gateway/event-persistence-writer.js";
+import type { NapcatImageMessageAnalyzer } from "../napcat/application/napcat-gateway/image-message-analyzer.js";
 import type { AgentMessageService } from "../agent/capabilities/messaging/application/agent-message.service.js";
 import type { IthomeService } from "../agent/capabilities/ithome/application/ithome.service.js";
 import type { StoryQueryService } from "../ops/application/story-query.service.js";
