+++ apps/server/src/app/server-runtime.ts
@@ -7,9 +7,9 @@ import { loadStaticConfig } from "../config/config.loader.js";
 import { createDbClient, type Database } from "../db/client.js";
 import { PrismaLlmChatCallDao } from "../llm/dao/impl/llm-chat-call.impl.dao.js";
 import { PrismaLogDao } from "../logger/dao/impl/log.impl.dao.js";
-import { PrismaImageAssetDao } from "../napcat/dao/impl/image-asset.impl.dao.js";
-import { PrismaNapcatEventDao } from "../napcat/dao/impl/napcat-event.impl.dao.js";
-import { PrismaNapcatQqMessageDao } from "../napcat/dao/impl/napcat-group-message.impl.dao.js";
+import { PrismaImageAssetDao } from "../napcat/infra/impl/image-asset.impl.dao.js";
+import { PrismaNapcatEventDao } from "../napcat/infra/impl/napcat-event.impl.dao.js";
+import { PrismaNapcatQqMessageDao } from "../napcat/infra/impl/napcat-group-message.impl.dao.js";
 import { BizError } from "../common/errors/biz-error.js";
 import { toHttpErrorResponse } from "../common/errors/http-error.js";
 import { AppLogHandler } from "../ops/http/app-log.handler.js";
@@ -46,8 +46,8 @@ import { TaskScheduler } from "../scheduler/application/task-scheduler.js";
 import { buildDataRetentionTasks } from "../scheduler/tasks/data-retention/data-retention-task.factory.js";
 import { SchedulerHandler } from "../scheduler/http/scheduler.handler.js";
 import { DefaultLlmChatCallQueryService } from "../ops/application/llm-chat-call-query.impl.service.js";
-import { NapcatEventPersistenceWriter } from "../napcat/service/napcat-gateway/event-persistence-writer.js";
-import { DefaultNapcatImageMessageAnalyzer } from "../napcat/service/napcat-gateway/image-message-analyzer.js";
+import { NapcatEventPersistenceWriter } from "../napcat/application/napcat-gateway/event-persistence-writer.js";
+import { DefaultNapcatImageMessageAnalyzer } from "../napcat/application/napcat-gateway/image-message-analyzer.js";
 import { HttpOssClient } from "../oss/oss-client.js";
 import { DefaultNapcatEventQueryService } from "../ops/application/napcat-event-query.impl.service.js";
 import { DefaultNapcatQqMessageQueryService } from "../ops/application/napcat-group-message-query.impl.service.js";
