+++ apps/server/src/napcat/application/napcat-gateway.impl.service.ts
@@ -6,7 +6,7 @@ import { AppLogger } from "../../logger/logger.js";
 import { type NapcatGatewayPersistenceWriter } from "./napcat-gateway/event-persistence-writer.js";
 import { NapcatGroupMessageProcessor } from "./napcat-gateway/group-message-processor.js";
 import type { NapcatImageMessageAnalyzer } from "./napcat-gateway/image-message-analyzer.js";
-import type { NapcatQqMessageDao } from "../dao/napcat-group-message.dao.js";
+import type { NapcatQqMessageDao } from "../infra/napcat-group-message.dao.js";
 import { NapcatGatewayInboundMessageRouter } from "./napcat-gateway/inbound-message-router.js";
 import { buildOutgoingMessageSegments, type WebSocketLike } from "./napcat-gateway/shared.js";
 import { NapcatGatewayTransport } from "./napcat-gateway/transport.js";
