+++ apps/server/src/napcat/application/napcat-gateway.service.ts
