+++ apps/server/src/napcat/application/napcat-gateway/qq-face-names.ts
