+++ apps/server/src/napcat/domain/napcat-segment.ts
