+++ apps/server/src/napcat/infra/image-asset.dao.ts
