+++ apps/server/src/ops/application/napcat-event-query.impl.service.ts
@@ -2,7 +2,7 @@ import {
   type NapcatEventListQuery,
   type NapcatEventListResponse,
 } from "@kagami/shared/schemas/napcat-event";
-import type { NapcatEventDao } from "../../napcat/dao/napcat-event.dao.js";
+import type { NapcatEventDao } from "../../napcat/infra/napcat-event.dao.js";
 import { mapNapcatEventList } from "../mappers/napcat-event.mapper.js";
 import type { NapcatEventQueryService } from "./napcat-event-query.service.js";
