+++ apps/server/src/ops/application/napcat-group-message-query.impl.service.ts
@@ -2,7 +2,7 @@ import {
   type NapcatQqMessageListQuery,
   type NapcatQqMessageListResponse,
 } from "@kagami/shared/schemas/napcat-group-message";
-import type { NapcatQqMessageDao } from "../../napcat/dao/napcat-group-message.dao.js";
+import type { NapcatQqMessageDao } from "../../napcat/infra/napcat-group-message.dao.js";
 import { mapNapcatQqMessageList } from "../mappers/napcat-group-message.mapper.js";
 import type { NapcatQqMessageQueryService } from "./napcat-group-message-query.service.js";
