+++ apps/server/test/agent/root-agent-session.test.ts
@@ -2,7 +2,7 @@ import { describe, expect, it } from "vitest";
 import { AppManager, type App } from "@kagami/agent-runtime";
 import { DefaultAgentContext } from "../../src/agent/runtime/context/default-agent-context.js";
 import { RootAgentSession } from "../../src/agent/runtime/root-agent/session/root-agent-session.js";
-import type { NapcatChatTarget } from "../../src/napcat/service/napcat-gateway.service.js";
+import type { NapcatChatTarget } from "../../src/napcat/application/napcat-gateway.service.js";
 import { initTestLoggerRuntime } from "../helpers/logger.js";
 
 initTestLoggerRuntime();
