+++ apps/server/test/dao/napcat-group-message.impl.dao.test.ts
@@ -1,7 +1,7 @@
 import { describe, expect, it, vi } from "vitest";
 import { type JsonValue } from "@kagami/shared/schemas/base";
 import type { Database } from "../../src/db/client.js";
-import { PrismaNapcatQqMessageDao } from "../../src/napcat/dao/impl/napcat-group-message.impl.dao.js";
+import { PrismaNapcatQqMessageDao } from "../../src/napcat/infra/impl/napcat-group-message.impl.dao.js";
 
 describe("PrismaNapcatQqMessageDao", () => {
   it("should persist structured message into message column", async () => {
