+++ apps/server/test/service/napcat-gateway.impl.service.test.ts
@@ -1,6 +1,6 @@
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
-import { NapcatEventPersistenceWriter } from "../../src/napcat/service/napcat-gateway/event-persistence-writer.js";
-import { DefaultNapcatGatewayService } from "../../src/napcat/service/napcat-gateway.impl.service.js";
+import { NapcatEventPersistenceWriter } from "../../src/napcat/application/napcat-gateway/event-persistence-writer.js";
+import { DefaultNapcatGatewayService } from "../../src/napcat/application/napcat-gateway.impl.service.js";
 import {
   FakeWebSocket,
   createAgentEventQueue,
