+++ apps/server/test/service/napcat-gateway.test-helper.ts
@@ -2,8 +2,8 @@ import { vi } from "vitest";
 import { type AgentEventQueue } from "../../src/agent/runtime/event/event.queue.js";
 import type { ConfigManager } from "../../src/config/config.manager.js";
 import type { Config } from "../../src/config/config.loader.js";
-import type { NapcatEventDao } from "../../src/napcat/dao/napcat-event.dao.js";
-import type { NapcatQqMessageDao } from "../../src/napcat/dao/napcat-group-message.dao.js";
+import type { NapcatEventDao } from "../../src/napcat/infra/napcat-event.dao.js";
+import type { NapcatQqMessageDao } from "../../src/napcat/infra/napcat-group-message.dao.js";
 import { initLoggerRuntime } from "../../src/logger/runtime.js";
 import type { LogEvent, LogSink } from "../../src/logger/types.js";
